+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "MusicFree",
-  "version": "0.4.3",
+  "version": "0.4.2",
   "private": true,
   "license": "AGPL",
   "author": {
